net: ks8851: Transmit packets in dedicated kthread
Transmission of packets is initiated in softirq context, hence cannot
sleep.  Because sleeping is required for use of the SPI bus, the ks8851
driver defers transmission to the work item ks8851_tx_work().

Communication with RevPi Gates should happen with realtime priority to
keep packet latency low.  That is cumbersome to implement with a work
item because it shares the worker thread with other work items and thus
would have to raise the worker thread's RT priority at the beginning
and lower it at the end of ks8851_tx_work().

Migrate the driver to a dedicated kthread for transmission of packets
to avoid this overhead by letting the kthread stay at RT priority for
consecutive packets sent to a RevPi Gate.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
[p.rosenberger: Port to v6.6.31]
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>