tpm: Silence self test error if uninitialized
If a TPM 2.0 chip has not been initialized by the platform firmware,
a message "A TPM error (256) occurred attempting the self test" is
emitted with KERN_ERR severity in addition to a KERN_INFO message
"starting up the TPM manually".  The former message is caused by
tpm_transmit_cmd() misinterpreting the (expected) TPM2_RC_INITIALIZE
response code as an error.

The error message is particularly inappropriate on TPMs which are not
used for integrity measurements but merely to store key material and
which are hence not touched by the firmware.  (In our case, a TPM
attached to a Raspberry Pi.)

Silence the error by adding the response code in question to the
existing whitelist of allowed response codes.

Signed-off-by: Lukas Wunner <lukas@wunner.de>