arm*/configs/revpi-*: Sync time from i2c rtc to system time on boot
As the RTC is on an i2c bus the i2c host drivers needs to be built in.
The same aplies to the RTC driver.

This seems to be fixed with:
f9b2a4d6a5f1 ("rtc: class: support hctosys from modular RTC drivers")
This commit might not be needed anymore but should only be removed
after proper testing and verification that it is indeed superfluos.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>