arm/configs/revpi-v7: 8192cu is unreliable on 4.9, blacklist it and unblacklist rtl8192cu
The Edimax EW-7811un USB WiFi adapter (vendor 0x7392, product 0x7811,
sold at the Revolution Pi online shop) is supported by two drivers in
the mainline kernel:

    drivers/net/wireless/realtek/rtl8xxxu
    drivers/net/wireless/realtek/rtlwifi/rtl8192cu

Additionally, the Foundation has patched in a third out-of-tree driver:

    drivers/net/wireless/realtek/rtl8192cu (module name: 8192cu.ko)

The rationale according to the commit message was that back in 2012,
8192cu was widely used and newer than rtl8192cu. The Foundation's
kernel only enables rtl8192cu and 8192cu, but blacklists the former
as well as rtl8xxxu.  8192cu is thus the default.

With 8192cu on 4.9 kernels, after bringing the Wifi interface down and
back up, reassociation with the base station fails and several WARN
splats appear in dmesg.

Blacklist 8192cu and unblacklist rtl8192cu.  This one seems to work
fine with kernel 4.9.  Compared to rtl8xxxu, it causes fewer interrupts
and thus lower CPU load.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
[p.rosenberger: moved from kernelbakery in the linux tree]
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>