dts: revpi-core: pull down sniff pin
There is no resistor to pull down sniff pin A2. This leads to a false
detection of a device on the left side of the pi-bridge in case that no
devices are connected: in this case a high signal level is read from the
line which is interpreted as an existing device during the pi-bridge
detection procedure.

Fix this by using the internal pull down of the pin controller.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>