arm/config/revpi-*: Disable /dev/gpiomem rootless GPIO access via mmap()
The CONFIG_BCM2835_DEVGPIOMEM option rovides users with root-free access
to the GPIO registers on the 2835. Calling mmap(/dev/gpiomem) will map the
GPIO register page to the user's pointer.

As this provides a way for unprivileged useres to change any h/w pin
this is a direct threat for security and might possibly damage the
hardware. To access GPIOs gpiod should be used.

Disable this option on all RevPi kernels.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>