tpm: allocate struct for tpm_tis_phy_ops on heap
Do not use a static struct for the phy ops but allocate it in
tpm_tis_spi_driver_probe(), pass it to tpm_tis_spi_probe and set the proper
function pointers.

This is in preparation of the next step when we set up SLB9670 specific
functions to assert and deassert the TPM chips reset.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>