dts/revpi-*: Switch to dwc2 USB driver
With the 5.10 kernel we see issues with the USB based network interface.
Some times the USB interface seems to be stalled and communication is
susspended. The received packages are processed after some time. The
delay can go up to several 10th seconds.
The issue seems to be with the usb_otg driver in combination with
piControl. Balancing the priorities (increase the dwc-otg irq threads
priorities or decreasing the piControl threads priorities) reduces the
problems, but doesn't solve them. The dwc2 driver has been developed
further while the development of the dwc-otg driver stagnated. So it
seems that the dwc2 driver doesn't suffer from these issues. So
switching to the dwc2 driver seems the obvious solution.

The issue can reproduced by running iperf in receive mode (iperf3 -R -c
...or iperf -s ...). Pings to the target might get lost and will be
received with a huge delay (up to 30 seconds and more). Also the
bandwith will be reduced some times. It might even drop to zero.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>