dts/compact: Switch to dwc2 USB driver
With the 5.10 kernel we see issues with the USB based network interface.
Some times the USB interface seems to be stalled and communication is
susspended. The received packages are processed after some time. The
delay can go up to several 10th seconds.
The issue seems to be with the usb_otg driver in combination with the
piControl io and ain threads. Balancing the priorities (increase the
dwc-otg irq threads priorities or decreasing the piControl threads
priorities) reduces the problems, but doesn't solve them.
The dwc2 driver has been developed further while the development of the
dwc-otg driver stagnated. So it seems that the dwc2 driver doesn't
suffer from these issues. So switching to the dwc2 driver seems the
obvious solution.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>