pibridge-serdev: Adjust driver name and compatible string
The serdev does not implement the piBridge but only a part of it, namely
the RS485 communication. So make it compatible to "kunbus,pibridge-comm"
instead of "kunbus,pibridge". This also avoids a name collision with the
device tree node that represents the piBridge and its sniff pins and
already uses the compatible "kunbus,pibridge".

Also adjust the driver name accordingly.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>