pibridge-serdev: Instead of pragma directive use packed attribute
Instead of a pragma(1) use the more common packed attribute to make sure
that structs pibridge_pkthdr_gate and pibridge_pkthdr_io are packed. This
is more explicit and also eliminates the danger that other structures that
are added in future get accidently packet when defined below the pragma
directive.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>