ARM: dts: revpi-flat-s-2022: Use mainline driver for WiFi
We've just amended the mwifiex mainline driver to support IW416,
the chipset on which Maya WiFi modules are based.

So use the mainline driver instead of NXP's out-of-tree driver in the
RevPi Flat S 2022 devicetree overlay.

While at it, take advantage of the existing SD8787 power sequencing
driver in the mainline kernel.

Signed-off-by: Lukas Wunner <lukas@wunner.de>