rtc: pcf2127: Fix the watchdog timeout
The timeout property of the watchdog_device is meant to be in seconds.
The watchdog_set_timeout() gets and just calls the
pcf2127_wdt_set_timeout() function with the timeout.

Initialize the timeout with PCF2127_WD_VAL_DEFAULT, which is in seconds.
The only place we actually write the timeout to the regitser is in
pcf2127_wdt_ping(). Just calculate the correct register value there.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>