PCI: brcmstb: support for PERST# over gpio
The Broadcom STB handles REFCLK and PERST# de-/assert in pcie controller.
But some PCIe devices have need longer TPERST#-CLK delays. With this
patch it is possible to manually control the de-/assertion of the PERST#
and thus implement longer TPERST#-CLK delays.

This needs a h/w design wich connects the PERST# of to a GPIO rater the
PERST# signal of the STB controller.

For now implement a rather long (100ms) delay.

Signed-off-by: Philipp Rosenberger <code@iluminat23.org>