dts: connect4: Unify gpio mux settings for rs485 entry
As denoted in the documentation of the broadcom gpio driver in
brcm,bcm2835-gpio.txt, either a single value or one for each pin
in brcm,function is possible. If a single entry in brcm,function
is given, it will be applied to all pins in that node.

Unify the brcm,function for rs485 to use only one entry like it is done
in the other pinctrl entries in this devicetree overlay.

Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>