arm*/configs/revpi-*: Set performance cpufreq governor as default
When the powersave governor is in use on the CM3, CM3+ or CM4, the
core clock is reduced from 1200 to 600 MHz on idle.

Because the SPI clock is derived from the core clock, throughput of
SPI-attached ks8851 network chips will then be halved as well.

Avoid by defaulting to the performance governor, which pegs the core
clock at the maximum 1200 MHz.

The higher clockrate is desirable for RT tasks in general.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>