arm: dts: revpi-flat-s-2022: Add DSA switch ports
Add DSA ports to the switch node and set the switch status to "disabled"
for now.

Also reduce the max SPI frequency from 20000000 Hz to 2000000 Hz:
Although in non-DSA mode the data transmission works with a frequency of
20000000 Hz this is not the case for DSA mode.

Since we use a phy mode of "rgmii" for the CPU port, we have to also set
rx and tx delays, in this case we use 2 nsecs for both rx and tx.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>