package: don't generate debian/changelog by default
The debianisation that can be generated with "scripts/package/mkdebian"
is not meant to be kept after building the kernel so it creates a
debian/changelog that is overridden every time the script is called.
Because the Revolution Pi linux fork wants a maintained "debian/" branch
the changelog shouldn't always be overridden when updating the "debian/"
folder.

Instead, introduce a new option "-c" to "scripts/package/gen_revpi_pkg"
that allows forcefully generating a new debian/changelog if needed/wanted.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>