package: mkdebian: add missing dependencies to linux package
While "linux-base" is not actually strictly needed to run the linux
kernel, unfortunately the package "raspi-firmware", which is needed for
running the kernel on a Raspberry Pi, requires this package for its
"postinst" maintainer script.

"kmod" is added because Debian's linux package also includes this.

Add the packages "kmod" and "linux-base" as dependencies for the main
linux kernel package.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>