pibridge-serdev: clear uninitialized fields in gateway datagram
Before commit 1e6a25ebbf41 ("pibridge-serdev: send datagrams in one") the
packet header was allocated separately and the cleared via memset which
ensured that the header fields do not contain uninitialized data.
Now the header is allocated together with the data part but the memset
is missing, so for the "src" and "seq" fields that data is not initialized.
Fix this by setting these fields explicitly to 0.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
(cherry picked from commit cba0ef0b5b74ff0b185decf5fb73dafd34a2c7ca)