iio: adc: mcp320x: Use workaround at AIN failure
It may happen that the AIN gets into a state in which it keeps returning an
invalid input value. In this case both BIT 22 and BIT 23 (i.e. the invalid
combination of overrange and underrange) are set.
This erroneous state can occur after many times of successful conversion
but once it occurred the chip remains in the state and keeps returning
invalid values.
We assume that the reason could be a failed wakeup from the shutdown mode
which according to the datasheet the chip automatically enters after each
conversion.

There is already a workaourond in the probe() routine of the mcp230x driver
to forcely wakeup the chip from the shutdown mode.

Use the same workaround (which consists of executing two consecutive
conversion without any delay in between) also in case that the chip starts
returning erroneous values.

With the workaround applied the invalid values are not reported in the log
file any more.

It is still unclear though, why the chip gets into this failure state and
how the workaround fixes it.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>