dts: Add Connect 5 with CAN controller (MCP251863)
A 20MHz oscillator is used for OSC1 and OSC2 input. It is modled as
can_20mhz_osc fixed-clock. The SPI SCK must be must be less than or equal
to 0.85*(FSYSCLK/2). Thus a max spi freuenzy of 8.5 MHz is selected.

The CAN termination can be switched on by driving gpio 16 of the gpio
expander on the core board high.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>