package: debian: add missing Depends to linux-headers
The linux-headers package is missing a dependency to libc. Lintian
complains about this as well:

    W: linux-headers-6.6.0-revpi3-rpi-v8: undeclared-elf-prerequisites (libc.so.6) [usr/src/linux-headers-6.6.0-revpi3-rpi-v8/scripts/asn1_compiler]
    W: linux-headers-6.6.0-revpi3-rpi-v8: undeclared-elf-prerequisites (libc.so.6) [usr/src/linux-headers-6.6.0-revpi3-rpi-v8/scripts/basic/fixdep]
    W: linux-headers-6.6.0-revpi3-rpi-v8: undeclared-elf-prerequisites (libc.so.6) [usr/src/linux-headers-6.6.0-revpi3-rpi-v8/scripts/genksyms/genksyms]
    W: linux-headers-6.6.0-revpi3-rpi-v8: undeclared-elf-prerequisites (libc.so.6) [usr/src/linux-headers-6.6.0-revpi3-rpi-v8/scripts/kallsyms]
    W: linux-headers-6.6.0-revpi3-rpi-v8: undeclared-elf-prerequisites (libc.so.6) [usr/src/linux-headers-6.6.0-revpi3-rpi-v8/scripts/mod/modpost]
    W: linux-headers-6.6.0-revpi3-rpi-v8: undeclared-elf-prerequisites (libc.so.6) [usr/src/linux-headers-6.6.0-revpi3-rpi-v8/scripts/sorttable]

Fix this by introducing "dh_shlibdeps" into the build process. It
calculated shared library dependencies for packages and creates a
substvar entry for them, making them usable in "debian/control" by
specifying "${shlibs:Depends}" in the "Depends" field.

"dh_shlibdeps" detects the dependency on libc for the linux-headers
package and by specifying the substvar for the linux-headers package in
"debian/control" this dependency is correctly declared and lintian won't
complain about it anymore.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>