dts/connect5: Use RC MSI target instead of MSI-MIP
The MSI-MIP peripheral provides a maximum of 8 vectors. This is not
enough for more than one LAN7430 downstream device. Thus the remaining
two devices will run with legacy interrupts.

Fix this and revert to the MSI target in the RC, instead of using the
MSI-MIP peripheral. DT changes are taken from RPi's
pciex1-compat-pi5-overlay.dts.

While at it make sure that the pcie1 interface is activated instead of
rely on some bootloader magic.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>