pibridge-serdev: change evaluation order for gate frames
Fix the order in which a received gate frame is evaluated by checking first
if the header is not corrupted. If it is corrupted do not bother with
receiving the packet data and CRC, because the probability is high that
also the data length field contains corrupted data.

Also add two more checks for invalid response flag settings (i.e. both OK
and ERR flag is set or none of both is set).

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>