pibridge-serdev: change evaluation order for io frames
Fix the order in which a received io frame is evaluated by checking first
if the header is not corrupted. If it is corrupted do not bother with
receiving the packet data and CRC, because the probability is high that
also the data length field contains corrupted data.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
(cherry picked from commit 3e4d45934773a4420185d4927ac4400ca7292ce3)
Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>