dts/connect5: External RS485 term is active high
The RevPi Connect 5 has one external RS485 interface, which also has a
switchable termination resistor. This resistor is switched with an opto
mosfet which is active high and not active low. Fix it and use the
correct value, so the state which can be set with SER_RS485_TERMINATE_BUS
is not inverted anymore.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>