dts/revpi-flasher: add overlay
The revpi-flasher is a board for flashing multiple RevPis at once. It is
driven by a Compute Module 5, which is mainly connected to a NVM Express
for a SSD, a DSI connector for a touch display, 8 USB 3.0 ports for
RevPis, 8 switchable power connectors for RevPis. The board also comes
with a UART interface, 2 freely addressable 3-color LEDs, and an HDMI
port for external video output.

Display specific parts of the overlay are taken from
vc4-kms-dsi-ili9881-7inch-overlay.dts.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>