pibridge-serdev: introduce function to send and receive complete datagrams
Introduce function pibridge_req_gate_datagram() which can be used to send
a complete datagram which is provided by the caller. The response is stored
in a response datagram, if available.
This makes it possible to have control over the header attributes like the
sequence number. It also allows to create invalid datagrams which can be
used to test the RevPi modules connected to the PiBridge.

Since a whole datagram also includes the header, make the gate datagram
header available by moving it from pibridge.h to pibridge_comm.h.
Also provide the structure for a complete datagram in this header file.

(cherry picked from commit 2eb00e638c7e8ca826850622e465ad7cf0e01322)
Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>