pibridge-serdev: increase internal fifo size to 512 byte
The current fifo size of 128 bytes is not enough for gateway packets that
carry the max payload of 256 bytes. Since those packets are sent by the new
mechanism in the firmware to retrieve debugging in tracing info, increase
the fifo size.
Make it 512 byte to be on the safe side.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>