pibridge-serdev: for gate datagrams apply timeout also for data and CRC
The timeout passed to pibridge_req_gate_tmt() is only applied to the
reception of the first bytes (i.e. the gate datagram header).
However especially in case of large data parts (up to 255 bytes) the
default timeout used by function pibridge_recv() may be too short.
So switch to pibridge_recv_timeout() for both data and CRC to make the use
of longer timeouts also for these parts possible.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>