pibridge-serdev: extend API by pibridge instance
Extend the API functions provided by the pibridge driver by a reference
to the driver. The reference itself can be retrieved by the new API
function pibridge_get().

Function pibridge_get() returns NULL as long as the probe() function was
not called. This can be the case if the serdev driver has not yet been
enumerated by the kernel.

The requirement for the pibridge driver reference prevents API users
from calling the API functions without the pibridge driver being
properly initialized.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>